separated ui specific code
